+++ miniconf/src/lib.rs
@@ -1,6 +1,6 @@
-#![cfg_attr(not(any(test, doctest, feature = "std")), no_std)]
+#![cfg_attr(not(any(test, doctest)), no_std)]
 #![cfg_attr(feature = "json-core", doc = include_str!("../README.md"))]
-#![cfg_attr(not(feature = "json-core"), doc = "Miniconf")]
+#![cfg_attr(not(feature = "json-core"), doc = "miniconf")]
 #![deny(rust_2018_compatibility)]
 #![deny(rust_2018_idioms)]
 #![warn(missing_docs)]