+++ miniconf/tests/iter.rs
@@ -1,33 +1,42 @@
-#![cfg(all(feature = "json-core", feature = "derive"))]
-
-use miniconf::{JsonCoreSlash, Tree, TreeKey};
+use miniconf::{JsonCoreSlash, PathIter, Tree, TreeKey};
 
 #[derive(Tree, Default, PartialEq, Debug)]
 struct Inner {
-    inner: f32,
+    inner: bool,
 }
 
 #[derive(Tree, Default, PartialEq, Debug)]
 struct Settings {
-    a: f32,
-    b: i32,
+    #[tree(depth = 1)]
+    b: [bool; 2],
     #[tree(depth = 1)]
     c: Inner,
+    #[tree(depth = 2)]
+    d: [Inner; 1],
+    a: bool,
 }
 
 #[test]
 fn struct_iter() {
-    let mut paths = ["/a", "/b", "/c/inner"].into_iter();
-    for (have, expect) in Settings::iter_paths::<String>("/").count().zip(&mut paths) {
-        assert_eq!(have.unwrap(), expect);
-    }
-    // Ensure that all fields were iterated.
-    assert_eq!(paths.next(), None);
+    assert_eq!(
+        Settings::iter_paths::<String>("/")
+            .count()
+            .map(|p| p.unwrap())
+            .collect::<Vec<_>>(),
+        ["/b/0", "/b/1", "/c/inner", "/d/0/inner", "/a"]
+    );
 }
 
 #[test]
 fn struct_iter_indices() {
-    let mut paths = [([0, 0], 1), ([1, 0], 1), ([2, 0], 2)].into_iter();
+    let mut paths = [
+        ([0, 0, 0], 2),
+        ([0, 1, 0], 2),
+        ([1, 0, 0], 2),
+        ([2, 0, 0], 3),
+        ([3, 0, 0], 1),
+    ]
+    .into_iter();
     for (have, expect) in Settings::iter_indices().count().zip(&mut paths) {
         assert_eq!(have, expect);
     }
@@ -37,19 +46,6 @@ fn struct_iter_indices() {
 
 #[test]
 fn array_iter() {
-    #[derive(Tree, Copy, Clone, Default)]
-    struct I {
-        c: bool,
-    }
-
-    #[derive(Tree, Default)]
-    struct Settings {
-        #[tree(depth = 1)]
-        a: [bool; 2],
-        #[tree(depth = 2)]
-        b: [I; 3],
-    }
-
     let mut s = Settings::default();
 
     for field in Settings::iter_paths::<String>("/").count() {
@@ -60,6 +56,36 @@ fn array_iter() {
         assert_eq!(&buf[..len], b"true");
     }
 
-    assert!(s.a.iter().all(|x| *x));
-    assert!(s.b.iter().all(|i| i.c));
+    assert!(s.a);
+    assert!(s.b.iter().all(|x| *x));
+    assert!(s.c.inner);
+    assert!(s.d.iter().all(|i| i.inner));
+}
+
+#[test]
+fn short_iter() {
+    assert_eq!(
+        PathIter::<Settings, 3, String, 1>::new("/")
+            .map(|p| p.unwrap())
+            .collect::<Vec<_>>(),
+        ["/a"]
+    );
+
+    assert!(PathIter::<Settings, 3, String, 0>::new("/")
+        .next()
+        .is_none());
+}
+
+#[test]
+#[should_panic]
+fn panic_short_iter() {
+    PathIter::<Settings, 3, String, 1>::new("/").count();
+}
+
+#[test]
+#[should_panic]
+fn panic_started_iter() {
+    let mut it = Settings::iter_indices();
+    it.next();
+    it.count();
 }